requirements22.txt
altair==5.5.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.0.1
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
narwhals==2.10.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.3.0
polars==1.34.0
polars-runtime-32==1.34.0
protobuf==6.33.0
psutil==7.1.2
pyarrow==22.0.0
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
safetensors==0.6.2
scipy==1.16.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
timm==1.0.22
toml==0.10.2
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.221
ultralytics-thop==2.0.17
urllib3==2.5.0
